Revert "Rename 'fix-csv' to 'fix_csv' (#13)"
This reverts commit 67bd3eb2a1a6e44f5e5c57c5219cfb5adbb6c666.

Underscore is discouraged only in package names, not in module names,
in https://www.python.org/dev/peps/pep-0008/#package-and-module-names.
In any case, our change had switched *to* underscore, not away; if we
were going to take the advice we had thought the PEP was giving, we
would have called the thing "fixcsv" :-).  But let's just revert
entirely.  This isn't a package anyway; it's at most a module, and
hyphen is more common in filenames (and usually easier to type).