Implement pare in fix-csv
Santizing is taking a long time.  It doesn't make sense to sanitize
everything, then throw 99% of that away.  If you need to check the
whole corpus for a formatting error, remove the pare in wiki-refresh.

The big downside here is that all our old entry numbers are now
divided by 100, since we've set pare to 100.  Entry 1300 is now entry
13.