Handle the "sanitized-" prefix reliably
We've been parsing the CSV filename out of the csv2wiki-config file,
ever since commit eceb2c92e and friends, but we hadn't been accounting
for the "sanitized-" prefix.  The result was that we were actually
sanitizing the result of some legacy sanitization, and producing a
file named "sanitized-sanitized-100andchangeExport-all-judges.csv".

Fix this by parsing out the prefix in order to get the name of the
original raw CSV file, and by documenting the expectation that
csv2wiki-config names the sanitized file.