Rename column headers
This sort of reverses commit 83e630f1142.  Although column header
rename functionality is now offered by csv2wiki, the Macarthur
Foundation would like the renames in the cleaned-up spreadsheet, not
just in the wiki.  So here we do the renames again, although this time
it's a broader set of renames, and done in a more robust way.