Stop specifying CSV in config file
Since commit c97de7a582f in csv2wiki, the CSV file has been specified
as a command-line parameter rather than in the config file -- which
means that our fancy footwork in wiki-refresh to parse the CSV file
name out from the config file has been obsolete this whole time.

Instead, just name the raw CSV file in wiki-refresh.  If the CSV file
is not in the current directory, the user can either move it there or
make a temporary edit to wiki-refresh to point to it.