Detect csvkit more conventionally
Checking for a source distrib of csvkit locks the user into installing
in just one way, and a non-standard way at that.  Instead, we just see
if the utilities are available and if the version string works.  We
don't even check the number on the version string because older
versions don't actually support --version.  If it doesn't error, it's
new enough.

Also, I removed the virtualenv stuff in the script.  This isn't the
place to force the user to use virtualenv.  If the user wants to use
virtualenv, the right place to do it is *before* the user runs
wiki-refresh.  That way the user gets to set their environment, which
is the whole point of virtualenv.  In my case, for example, I just did
a `pip install csvkit` in my Macfound virtualenv.  It made the tools
available while that virtualenv was active, and that was good enough
to make it available to wiki-refresh.