Name delimiter & quotechar when not sniffing (#38)
In the csvjoin invocation for which CSV sniffing was disabled in
commit c8b45eaf, specify --delimiter and --quotechar explicitly.

This makes no difference right now, as the delimiter and quotechar we
pass are the defaults anyway, but it's more robust to name them.
While it seems unlikely that those defaults would ever change, we
might as well be future-proof, especially if not doing any sniffing.