Add small output for # of scrubbed cells.
The fix_cell method in compose-csvs is based on multiple levels
of data sets, and in the future may actually not be necessary.  This
adds a small output noting the number of cells that the scrubbing
actually did something to, so we can reevaluate if that number starts
being regularly zero.