Add TOC by geography.
The categories were not working sufficiently for macarthur, so this adds
the option to specify that we want to use TOCs, and link to them.  Added
as an option so that we can continue to use categories if we want.
Then, the first one was by geography, so that was added at the same time.