Add extra-output-csv argument.
This is based on a requirement that a basic version of the top 200
(without review data) is output for a third party.  This is done inline
with creating the other rows because of how the system is grown
organically.  It would make more sense to create an object representing
the data with different output points, but for now, this works well.