Rip out the old method of csv2wiki.
Before, we were uploading all the wiki pages using csv2wiki, after
generating them.  We are moving to a place where torque data handles all
these, and so we need to clean up the code as a foundation for the new
method.  Some things will get re-added, but this now creates a pure
spreadsheet that doesn't have wiki markup added.