Add TDC configuration files
With the torque-data config being moved to the wiki, and out of the
torque-data app, we needed a place to put the base information that was
being used/generated.  Even though these aren't strictly needed for
running wikis, this is a good place for them because people deploying
new datasets can use them as a reference.