Remove old files from 2016 100&Change competition upload.
With LFC in 2019 in full swing, the old files that worked on the 2016
dataset no longer work, nor are referenced from any of the new versions
of the ETL pipeline scripts.  After evaluating, there was no real
information in them that needed to be preserved for future use, so could
be safely removed.