Add 2017 100 And Change ETL pipeline.
This took the pipeline that existed in cb5c587 before we moved to the
2020 competition, restores it into the proper subdirectory, and
modernizes it for the new multi competition torque world.

This included splitting it into two projects, both 100Change2017 and
the secondary view on that aday in 100Change2017Partners, as well
as removing a bunch of old files.

This adds both the ansible files and the etl pipelines for those
competitions.