Build ansible inventories
Ansible inventories for competitions should now live in this repository
because even though they are specific to OTS's setup, they can be public
information for the most part, and should be kept with the etl pipeline
and ansible scripts.

Inventories are created for all environments competitions are currently
deployed on, and also for local deployments.