Move mwlib ansible install into its own directory
Because there's a many to one relationship between torque-site
deployments and mwlib (as many of them can have book printing), we
really want to either have all of them deploy it or none of them and
require the sysadmin to manage deploying it as needed.

The decision to move it to its own is the second option, and has the
added benefit of allowing deployments of individual wikis move faster.