Remove tech debt: torque_dir
While the project was in limbo, torque_dir referred to the location for
your checkout of the torque project which had a lot of custom extensions
and the torquedata server itself.  This change removes that and gets the
code directly from git in order to be more maintainable and reproducable
across developers and time.