Move simplesaml ansible install into its own directory
Because there's a many to one relationship between torque-site
deployments and simplesaml, we really want to either have all of
them deploy it or none of them and require the sysadmin to
manage deploying it as needed.

The decision to move it to its own is the second option, and has the
added benefit of allowing deployments of individual wikis move faster.

This requires adding ansible roles for competitions that installs
a metadata for a given install that's given to us by macarthur.

There's a complementary change to this move in opass that puts
the salts into a top level simplesamlsalt pass