Restore top-level README.md
This reverts part of commit 8b3341d1538, which overly-eagerly moved
the entire top-level README.md into the 100Change2020/etl/ subdir.
While it is true that most of that README.md was specific to that
competition's ETL process, the first section was a high-level
explanation of this overall repository, so that's what I've restored.