Add prototype budget line item extractor.
For EO2020, one of the fields accessible in the PDF but not in the csv
was budget line items.  This code was the first attempt at extracting
it, but ran into the fact that the pdf's were not getting converted to
html by xpdf in a regular enough fashion to be usable.