Add lfcrecommendations, update thematic areas to have description
This accidentally got two disparate changes in one, but they are:

Adding the lfc recommendations, which was a small spreadsheet with
just a classification of 25 proposals based on a document LFC produced
internally.

Also update thematic areas to have descriptions included.  This
coincided with some other changes requiring us to not overwrite the
pages on the wiki by default.