Add supplemental attachments for top 13 proposals
Though these aren't actually the "top 13", but rather 13 that the board
members were interested in.

Requires putting the attachments in the right place, and then creating
a new TOC and TOC template for these to add to a page they specified.

Also kept the list of what exactly are the top13 in svn to keep it
private.