User username attribute from okta for local usernames.
Before, we were using email because the username they were sending over
was the email.  But they updated their okta configuration to pass over a
specific username they'd like to see relating to this application, and
we want to use it.

This is related to some users having very awkward number based email
addresses that were set up, and would like their username in the wiki to
be a different, more identifiable, name.