Change supplemental toc to use list form
We're replacing the complex supplemental TOC with a more configurable
template that lives on the wiki.  Before, we were stretching the meaning
of a table of contents and the work being done by torque, and having
them do more than stictly necessary.  Now that configuration is in the
right place.