Add two new user groups to 100Change2017
Add the "LFC Research Partners" group and the "LFC Evaluators" group
to 100Change2017.

I'm not sure about this change, because 100Change2017 seemed to differ
from 100Change2020 in more ways than just not having these two groups.
For example, the entire collection of CSS-related stanzas in main.yml
was missing in 100Change2017.  This commit adds those stanzas, along
with the two new groups being listed in the 'loop:' specification, but
what about "BoardMembersTorque" and "BoardMembers" from 100Change2020
in the "Install board group css" stanza -- should they be added in
100Change2017 as well?

Therefore, please take this as "Karl trying to save someone else time
by doing some of the work in advance" rather than "Karl thinks this
change is correct".