Rework recent edit to avoid changing anchor link
In commit d03bbe42eeed I added the word "roles" to a section title in
order to make the section findable via a search for that word.  This
had the unintentional side effect of changing the canonical anchor
link for that section, from "https://..DESIGN.md#user-groups" to
"https://.../DESIGN.md#user-groups--roles".  But the original link is
one that we've shared with others, including with Lever for Change, so
we should try to keep it stable.  This commit restores the original
section title (and thus the original anchor link), and moves the word
"roles" into the first paragraph so we can still find it in a search.