Deploy new DemoView with edit mode enabled
This change required a few other changes to come in.  DemoView needed to
update to point to a different torquedata server, which requried
torquedata to have a new ansible inventory allocated (demo).  This also
required demoview and torquedata to pull in new changes that allowed
torquedata supervisor startup to set the port, and the torquedataconnect
extension to read it.