Add Central ETL Pipeline, version 1
This adds a centralized, refactored, etl pipeline that all competitions
can depend on.  This prevents the rampant copying and pasting of etl
pipelines as new competitions are created, and prevents the coderot from
modifying later etl pipelines that don't make it back into the original
ones.

Version 1 because this is based on the Climate2030 pipeline that doesn't
have a number of features that older, more complete pipelines have.
Those will get added incrementally.

Issue #53: Refactor ETL code to create central utilities