Update EO2020 to use the refactored etl pipeline
This couldn't really be a diff, especially on compose-csvs/upload, because
the change was too massive.  So those files were deleted and a new one was
put in their place, even though it has some of the options processing
those did.

This was a more interesting case because a lot of the work in the EO2020
pipeline was unique to EO2020, which required the compose-and-upload
file to have a few overridden classes so that the main etl pipeline
module doesn't get cluttered with one-offs.

Issue #53: Refactor ETL code to create central utilities