Update 100Change2017 to use the refactored etl pipeline
This is the latest transformation of the old pipeline, which removes all
remnants of csv2wiki.  The fix-csvs is closer to how it actually was
originally, and all the torque work is over in compose-and-upload now.
It's named that mainly to be in alignment with the other competitions,
even though there's no composing going on.

Issue #53: Refactor ETL code to create central utilities