Remove outdated script comments
Each of the compose files came from an original script that was very
well documented, and that documentation transformed and moved into
the top level README's and whatnot, where it should be.  It also became
out of date and confusing.

From review on #53: Refactor ETL code to create central utilities