Remove arbitrary default rank attribute, fix sorting
Now uses the proper python sorting with comparison functions on the
Attachment object, with a default rank of None that gets ordered to the
bottom.  The old rank of 99 was both broken (the sort was written
improperly), and weirdly a magical number.

From review #77: Refactor ETL code to create central utilities