Use the new refactored ansible with 100Change2017
With the ansible configuration centralized into a top level torque-sites
directory (roles), the majority of the tasks main.yml and supporting
files can be removed with the roles being added to the top level
playbook.

This required a few changes to the base pipeline:
* Add composer to mediawiki install
* Composer version update
* Add postgres base package
* Add LFC Evaluators as a known SAML group
* Add pre-req installation for mediawiki

Then removing all the old stuff from 100Change2017 and updating
the yml file to reference the general stuff.