Change the default login mechanism to Okta
Based on conversations with MacFound, the base way we want to log into
competitions should be okta, with a locallogin.php option for special
cases.  That means that before, where competitions that defaulted to
okta were the exception, and had the overriding configuration in their
ansible tasks now need to do nothing.  Competitions that need to always
have the local login option (due to legacy reasons) enabled are now the
exception, and have those exceptions in their ansible files.