Move thirdparty libs into roles directory that matches them
Before the ansible refactor, we were copying and pasting a lot of
ansible configs.  Instead of copying and pasting the large tar balls
that went with those deployments, they all lived in a central
'thirdparty' directory.  However, that creates confusing ansible
playbooks and directory structure, and it's much simpler to have things
co-located.  This resolves that change.