Allow new columns to be lists
When adding a new column for attachments, there can be multiple matches
for the regex.  In that case, multiple items get added for a new column.
This change allows that column to be specified as a list column so that
the spreadsheet looks correct.

There's also a small bug fix where column_name was getting assigned to
name for new column attachments, which broke sorting.