Update competition templates
This continues the change where we replace template placeholder values
in ansible variable templates with a placeholder format that can work
with `envsubst` to insert environment variables.

Since this is just a modification to the template placeholders this should
have no impact on installation steps (as those values would be replaced
by hand anyway).  All this does is make it possible for a script to
populate a template more easily.

Issue #85