Add mysql to mediawiki deps
Mediawiki sometimes needs the php-mysql task to run.  It doesn't really
hurt to proactively include it here (we already do so for psql).

The alternative would be to remove database dependency installations
from the mediawiki task and require that the calling role determine
which dependency to include, but that is a larger scope concept and not
necessarily something we would prefer anyway.

Issue #93