Add support for gpg passphrase to 100Change2020
The 100Change2020 deploy script supports the downloading and decrypting
of key competition data before running the ingestion. This could not be
scripted in non-interactive environments because there was no support to
pass in a GPG passphrase.

This adds a new `-g` flag to deploy, which accepts a passphrase.

We will likely want to support this flag in other competition deploy
scripts, and it may be valuable to consider creating a set of common
utility methods imported across the various deploy scripts to avoid DRY
violation.

Issue #93