Add autoconfig conditional
There was a potential bug where simplesaml autoconfig would populate
with an empty '' => '' if the okta variables were not set.

This adds a conditional, skipping that step if there is no value
present.  There are still potentials for error in the event that
simplesaml is never provided any value, however, as at least one
authentication endpoint must be defined for simplesaml to work.

Issue #93