Fix word wrap / page width
There was an issue with long URLs breaking word wrap and the solution at
the time was to set a very wide min-width. This is not ideal for all
viewports, including printing, so we are moving to a word-wrap rule.

Because of flex box there is an additional improvement necessary, which
is that flex-boxes will specify a default min width of whatever the
longest element (before word breaks).  By specifying a smaller min-width
we prevent that auto calculation.

https://stackoverflow.com/questions/36150458/flex-item-overflows-container-due-to-long-word-even-after-using-word-wrap/36150638

Issue #120