Bring 100Change2020Demo up to date
It didn't get changes after the ansible refactor, and so failed to work
with the mediawiki upgrade correctly.  This brings it up to date with
the rest of the competitions, and corrects some other errors (like
referencing proposals instead of LFC100Change2020 in etl pipeline) so
that it once again works in production.