Separate json based competitions and csv based ones
There are competitions where we are not basing them on a spreadsheet,
but rather from the output of a different competition's run.  Those
competitions need to load up the data from the outputted json file,
rather than from a base spreadsheet.  So we subclass Competition to
allow either of those.