Remove LLIIA2020 geolocations
These were added for completeness originally, but are not being used and
result in many more calls to the geocoder on ETL run.

Taking them out will speed up ETL; if they are needed later (and / or
when local caching of geo data is implemented) we will add them back.