Allow geocoding in debug
Geocoding is an expensive process, and sometimes we want to be able to
run local ETL scripts that generate geo data but don't actually run
against a geocoder.  The `debug` flag enables this functionality.

This is intended to only be used for local testing, and the flag may
disappear in future.

Issue #139