Preparing for the LFC group rebuild
This change keeps both the old groups, and adds the new ones, so that
LFC can change their okta set up and nothing is lost.

In addition, groups that aren't specified now default to "LfcOkta" to
make the user groups more consuable when looking at the user list.

The change from DecisionMakers.css to Restricted.css is because now it's
for most everyone who has a restricted UI (non admins).