Change sidebar for views
This required a torque update to get the code out of there.  In
addition, the views are no longer required for the different
competitions, only globalview.  That sidebar code is moved into this
repository as a specially loaded php file that adds it.

A small change was needed to the simplebook_tdc to pull the view
correctly as the function was removed.