Support tab overrides
This updates our tab hook logic to allow a given torque site to define
a series of custom tab overrides that can be specified on a per-page
basis.

The overrides are defined using a new $wgLFCExtraTabOverrides global,
which is a dict where keys are the name of the tab override and the
value is the value that we want to be used for $wgLFCExtraTabs.

For example:

$wgLFCExtraTabOverrides = [
  "FINALIST" => [
    ["Factsheet for ", "Factsheet", "main"],
    ["Prospectus for ", "Prospectus", "prospectus"],
  ]
]

To invoke the FINALIST tabset, the wiki page needs to include:

<!-- TABS: FINALIST -->

Issue #159