Supplement and clean up Sustainable Development Goals
For some competitions, this is a static field as there's a single SDG
for that competition.  For 2017, it's a corrections file with specified
per proposal.  Then update the tocs to display them with the new format
(<number>: <title>).  This required adding the ability to hide the auto
generated TOC numbering from mediawiki.